Pretvori u profakturu

+++ src/TD.Web/TD.Web.Public/TD.Web.Public.Contrats/Dtos/Orders/OrderGetSingleDto.cs
@@ -6,7 +6,7 @@ public class OrderGetSingleDto
     public string? OneTimeHash { get; set; }
     public int? KomercijalnoVrDok { get; set; }
     public int? KomercijalnoBrDok { get; set; }
-    public int? StoreId { get; set; }
+    public long? StoreId { get; set; }
     public DateTime? CheckedOutAt { get; set; }
     public required string Status { get; set; }
     public required int StatusId { get; set; }
